refactor: 拆分大文件（registry/player/breakthrough/tribulation）
- registry.ts (309行) → registry/ 目录 5 文件（stores/dlc/queries/event-engine/index）
- player.ts (258行) → player/ 目录 4 文件（types/create/stats/index）
- breakthrough.ts (206行) → breakthrough/ 目录 3 文件（status/attempt/index）
- tribulation.ts (197行) → tribulation/ 目录 3 文件（wave-combat/run/index）

所有新文件 ≤ 100 行，barrel re-export 保持向后兼容
