+++ src/game/tribulation/index.ts
@@ -1,0 +1,5 @@
+// tribulation/index.ts
+export type { TribulationResult } from './run';
+export { runTribulation } from './run';
+export type { WaveResult } from './wave-combat';
+export { fightWave } from './wave-combat';
