+++ src/hooks/useGameEngine.ts
@@ -1,24 +1,16 @@
 // ============================================================
-// useGameEngine.ts — 游戏核心引擎 Hook
-// A-1~A-5 + B-1~B-5 全部系统的状态管理 + 存档
+// useGameEngine.ts — 游戏核心引擎 Hook（状态管理 + 存档 + 编排）
+// 行为逻辑已拆分至 useCoreActions.ts / useSystemActions.ts
 // ============================================================
 
 import { useState, useCallback, useEffect } from 'react';
-import { createPlayer, recalcStats, getNextRealm, getSpiritRootGrade } from '../game/player';
+import { createPlayer, getSpiritRootGrade } from '../game/player';
 import type { Player } from '../game/player';
-import { REALMS, ACTION_COSTS, MONSTERS, BASE_CULTIVATE_EXP } from '../game/data';
-import { runCombat } from '../game/combat';
-import { registerCoreEvents, triggerExploreEvent, triggerDailyEvent } from '../game/events';
-import { useItem as inventoryUseItem, addItem } from '../game/inventory';
-import { getItemDef } from '../game/registry';
-import { performAlchemy } from '../game/alchemy';
-import { equipItem, unequipItem } from '../game/equipment';
-import { buyItem, sellItem } from '../game/shop';
-import { performSmithing } from '../game/smithing';
-import { attemptBreakthrough as attemptBreakthroughFn } from '../game/breakthrough';
-import { runTribulation as runTribulationFn } from '../game/tribulation';
-import type { EquipSlot } from '../game/registry';
+import { REALMS, ACTION_COSTS } from '../game/data';
+import { registerCoreEvents, triggerDailyEvent } from '../game/events';
 import type { LogCategory } from './useGameLog';
+import { useCoreActions } from './useCoreActions';
+import { useSystemActions } from './useSystemActions';
 
 // 注册核心事件（模块加载时执行一次）
 registerCoreEvents();
@@ -102,7 +94,7 @@ export function useGameEngine(addLog: (msg: string, category?: LogCategory) => v
       }
     }
     return updated;
-  }, [addLog]);
+  }, [addLog, gameOver]);
 
   // ── 精力检查 ──
   const canAct = useCallback((actionKey: string): boolean => {
@@ -111,217 +103,15 @@ export function useGameEngine(addLog: (msg: string, category?: LogCategory) => v
     return player.stamina >= cost.stamina;
   }, [player, gameOver]);
 
-  // ── A-2: 修炼 ──
-  const cultivate = useCallback(() => {
-    if (!canAct('cultivate')) {
-      addLog('⚠️ 精力不足，请先休息！', 'system');
-      return;
-    }
-    setPlayer(prev => {
-      if (!prev) return prev;
-      let p: Player = { ...prev };
-      const cost = ACTION_COSTS.cultivate;
+  // ── 子 Hook：核心行为（修炼/战斗/探索/休息）──
+  const { cultivate, fight, explore, rest } = useCoreActions({
+    addLog, addLogs, setPlayer, advanceTime, canAct,
+  });
 
-      // 消耗精力
-      p.stamina -= cost.stamina;
-
-      // 经验计算：base × 悟性加成 × 灵根加成 × 心情加成
-      const compBonus = 1 + p.comprehension / 50;
-      const rootGrade = getSpiritRootGrade(p.aptitudes);
-      const moodBonus = 0.5 + (p.mood / 100);
-      const expGain = Math.floor(BASE_CULTIVATE_EXP * compBonus * rootGrade.multiplier * moodBonus);
-
-      p.exp += expGain;
-
-      // 追踪连续修炼
-      p.tracking = { ...p.tracking, consecutiveCultivates: p.tracking.consecutiveCultivates + 1, consecutiveRests: 0 };
-
-      // 推进时间
-      p = advanceTime(p, 'cultivate');
-
-      addLog(`🧘 修炼一次，获得 ${expGain} 修为。（悟性×${compBonus.toFixed(1)} 灵根×${rootGrade.multiplier} 心情×${moodBonus.toFixed(1)}）`, 'system');
-      return p;
-    });
-  }, [canAct, advanceTime, addLog]);
-
-  // ── A-3: 战斗 ──
-  const fight = useCallback(() => {
-    if (!canAct('combat')) {
-      addLog('⚠️ 精力不足，请先休息！', 'system');
-      return;
-    }
-
-    setPlayer(prev => {
-      if (!prev) return prev;
-      let p: Player = { ...prev };
-      const cost = ACTION_COSTS.combat;
-      p.stamina -= cost.stamina;
-
-      // 选择同境界或 ±1 的随机妖兽
-      const eligible = MONSTERS.filter(m =>
-        m.realmIndex >= p.realmIndex - 1 && m.realmIndex <= p.realmIndex
-      );
-      if (eligible.length === 0) {
-        addLog('🔍 四周平静，没有发现妖兽。', 'combat');
-        return p;
-      }
-      const monster = eligible[Math.floor(Math.random() * eligible.length)];
-
-      const result = runCombat(p, monster);
-      addLogs(result.logs, 'combat');
-
-      p.hp = result.playerHpLeft;
-      p.exp += result.expGained;
-      p.gold += result.goldGained;
-
-      if (result.winner === 'player') {
-        p.tracking = { ...p.tracking, killCount: p.tracking.killCount + 1, consecutiveRests: 0, consecutiveCultivates: 0 };
-        // 检测是否击败高境界敌人
-        if (monster.realmIndex > p.realmIndex) {
-          p.tracking = { ...p.tracking, defeatedHigherRealm: true };
-        }
-        // C-1: 战斗掉落物品（30% 獠牙，10% 妖丹）
-        if (Math.random() < 0.3) {
-          const { player: p2, added } = addItem(p, 'core:monster_fang', 1);
-          p = p2;
-          if (added > 0) addLog('🦴 获得 妖兽獠牙 ×1', 'combat');
-        }
-        if (Math.random() < 0.1) {
-          const { player: p2, added } = addItem(p, 'core:monster_core', 1);
-          p = p2;
-          if (added > 0) addLog('💎 获得 妖丹 ×1', 'combat');
-        }
-        // T0014: 战斗掉落装备（5% 按境界）
-        if (Math.random() < 0.05) {
-          const equipDrops: Record<number, string[]> = {
-            0: ['core:iron_sword', 'core:cloth_hat', 'core:linen_robe', 'core:straw_shoes', 'core:jade_pendant'],
-            1: ['core:spirit_sword', 'core:spirit_crown', 'core:spirit_robe', 'core:wind_boots', 'core:spirit_ring'],
-            2: ['core:flame_blade', 'core:ice_helm', 'core:golden_armor', 'core:shadow_boots', 'core:dragon_amulet'],
-            3: ['core:thunder_spear'],
-          };
-          const tier = Math.min(monster.realmIndex, 3);
-          const pool = equipDrops[tier] || equipDrops[0];
-          const dropId = pool[Math.floor(Math.random() * pool.length)];
-          const { player: p2, added } = addItem(p, dropId, 1);
-          if (added > 0) {
-            p = p2;
-            const eDef = getItemDef(dropId);
-            addLog(`⚔️ 获得装备 ${eDef?.name ?? dropId}！`, 'combat');
-          }
-        }
-      } else if (result.winner === 'monster') {
-        // 战斗失败：健康 -20，重伤恢复少量 HP
-        p.health = Math.max(0, p.health - 20);
-        p.hp = Math.max(1, Math.floor(p.maxHp * 0.1));
-        if (p.hp < p.maxHp * 0.1) {
-          p.tracking = { ...p.tracking, hasBeenBelow10Hp: true };
-        }
-      }
-
-      p = advanceTime(p, 'combat');
-      return p;
-    });
-  }, [canAct, advanceTime, addLog, addLogs]);
-
-  // ── 探索 ──
-  const explore = useCallback(() => {
-    if (!canAct('explore')) {
-      addLog('⚠️ 精力不足，请先休息！', 'system');
-      return;
-    }
-
-    setPlayer(prev => {
-      if (!prev) return prev;
-      let p: Player = { ...prev };
-      const cost = ACTION_COSTS.explore;
-      p.stamina -= cost.stamina;
-      p.tracking = { ...p.tracking, consecutiveRests: 0, consecutiveCultivates: 0 };
-
-      const { player: updated, message } = triggerExploreEvent(p);
-      p = { ...updated };
-      addLog(message, message.includes('【奇遇】') ? 'adventure' : 'explore');
-
-      // C-1: 探索随机拾取物品
-      const exploreLoot: [string, number][] = [
-        ['core:iron_ore', 0.15],
-        ['core:spirit_stone_shard', 0.20],
-        ['core:herb_lingzhi', 0.08],
-        ['core:herb_snow_lotus', 0.03],
-        ['core:jade_slip', 0.05],
-        ['core:hp_pill', 0.10],
-        ['core:spirit_water', 0.06],
-        ['core:map_fragment', 0.02],
-      ];
-      for (const [itemId, chance] of exploreLoot) {
-        if (Math.random() < chance) {
-          const { player: p2, added } = addItem(p, itemId, 1);
-          if (added > 0) {
-            p = p2;
-            const def = getItemDef(itemId);
-            addLog(`🎁 拾取 ${def?.name ?? itemId} ×1`, 'explore');
-          }
-          break; // 每次探索最多捡一个物品
-        }
-      }
-
-      p = advanceTime(p, 'explore');
-      return p;
-    });
-  }, [canAct, advanceTime, addLog]);
-
-  // ── 休息 ──
-  const rest = useCallback(() => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      let p: Player = { ...prev };
-      // 回复 30% 精力、少量 HP/MP
-      const staminaRecover = Math.floor(p.maxStamina * 0.3);
-      p.stamina = Math.min(p.maxStamina, p.stamina + staminaRecover);
-      p.hp = Math.min(p.maxHp, p.hp + Math.floor(p.maxHp * 0.1));
-      p.mp = Math.min(p.maxMp, p.mp + Math.floor(p.maxMp * 0.1));
-      p.health = Math.min(100, p.health + 5);
-      p.mood = Math.min(100, p.mood + 3);
-
-      p.tracking = { ...p.tracking, consecutiveRests: p.tracking.consecutiveRests + 1, consecutiveCultivates: 0 };
-
-      p = advanceTime(p, 'rest');
-      addLog(`💤 休息片刻，恢复 ${staminaRecover} 精力，HP/MP/健康/心情少量恢复。`, 'system');
-      return p;
-    });
-  }, [advanceTime, addLog]);
-
-  // ── T0029: 突破系统重构 ──
-  const breakthrough = useCallback(() => {
-    if (!player) return;
-
-    setPlayer(prev => {
-      if (!prev) return prev;
-
-      // 先尝试普通突破
-      const btResult = attemptBreakthroughFn(prev);
-      for (const log of btResult.logs) {
-        addLog(log, 'system');
-      }
-
-      if (btResult.triggerTribulation) {
-        // 需要渡劫
-        const tribResult = runTribulationFn(btResult.player);
-        for (const log of tribResult.logs) {
-          addLog(log, 'system');
-        }
-
-        if (!tribResult.success && tribResult.player.hp <= 0) {
-          // 渡劫失败且判定为死亡
-          setGameOver(true);
-          setGameOverReason('渡劫失败，形神俱灭！');
-        }
-
-        return tribResult.player;
-      }
-
-      return btResult.player;
-    });
-  }, [player, addLog]);
+  // ── 子 Hook：系统行为（炼丹/炼器/装备/商店/背包/突破）──
+  const { useItem, craft, equip, unequip, buy, sell, smith, breakthrough } = useSystemActions({
+    player, addLog, setPlayer, setGameOver, setGameOverReason,
+  });
 
   // ── 删档 ──
   const deleteSave = useCallback(() => {
@@ -329,74 +119,6 @@ export function useGameEngine(addLog: (msg: string, category?: LogCategory) => v
     setPlayer(null);
     setGameOver(false);
     addLog('🗑️ 存档已删除。', 'system');
-  }, [addLog]);
-
-  // ── C-1: 使用物品 ──
-  const useItemAction = useCallback((itemId: string) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = inventoryUseItem(prev, itemId);
-      addLog(result.message, result.success ? 'system' : 'system');
-      return result.player;
-    });
-  }, [addLog]);
-
-  // ── T0013: 炼丹 ──
-  const craft = useCallback((recipeId: string) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = performAlchemy(prev, recipeId);
-      addLog(result.message, 'system');
-      return result.player;
-    });
-  }, [addLog]);
-
-  // ── T0014: 装备 ──
-  const equip = useCallback((equipId: string) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = equipItem(prev, equipId);
-      addLog(result.message, 'system');
-      return result.player;
-    });
-  }, [addLog]);
-
-  const unequip = useCallback((slot: EquipSlot) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = unequipItem(prev, slot);
-      addLog(result.message, 'system');
-      return result.player;
-    });
-  }, [addLog]);
-
-  // ── T0015: 商店 ──
-  const buy = useCallback((itemId: string) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = buyItem(prev, itemId);
-      addLog(result.message, 'system');
-      return result.player;
-    });
-  }, [addLog]);
-
-  const sell = useCallback((itemId: string) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = sellItem(prev, itemId);
-      addLog(result.message, 'system');
-      return result.player;
-    });
-  }, [addLog]);
-
-  // ── T0016: 炼器 ──
-  const smith = useCallback((recipeId: string) => {
-    setPlayer(prev => {
-      if (!prev) return prev;
-      const result = performSmithing(prev, recipeId);
-      addLog(result.message, 'system');
-      return result.player;
-    });
   }, [addLog]);
 
   return {
@@ -412,7 +134,7 @@ export function useGameEngine(addLog: (msg: string, category?: LogCategory) => v
     breakthrough,
     deleteSave,
     canAct,
-    useItem: useItemAction,
+    useItem,
     craft,
     equip,
     unequip,
